requirements.txt
flask
psycopg2-binary
pymongo
bcrypt
PyJWT
python-dotenv
requests
